+++ src/components/Search/tabs.js
@@ -1,0 +1,10 @@
+export const TABS = [
+  {
+    index: 'Assets',
+    content: 'Assets'
+  },
+  {
+    index: 'Trends',
+    content: 'Trends'
+  }
+]
