+++ src/components/TimelineEventComments/types.js
@@ -1,0 +1,4 @@
+export const CommentTypes = {
+  TIMELINE_EVENT: 'TIMELINE_EVENT',
+  INSIGHT: 'INSIGHT'
+}
