+++ src/components/WalletLink/WalletLink.js
@@ -2,7 +2,7 @@ import React from 'react'
 import cx from 'classnames'
 import SmoothDropdownItem from './../SmoothDropdown/SmoothDropdownItem'
 import ViewBalanceDialog from './ViewBalanceDialog'
-import TransactionTableLabels from './TransactionTableLabels'
+import Address from './Address'
 import styles from './WalletLink.module.scss'
 
 const WalletLink = ({
@@ -49,42 +49,6 @@ const WalletLink = ({
       />
     )
   }
-}
-
-const EtherscanLink = ({
-  address = '',
-  isTx,
-  isExchange,
-  label,
-  isFull,
-  asLink = true,
-  children
-}) => {
-  const link = children || address
-  const addressShort = isFull
-    ? link
-    : link.slice(0, isExchange ? 7 : 16) + '...'
-  return (
-    <a
-      href={
-        asLink
-          ? `https://etherscan.io/${isTx ? 'tx' : 'address'}/${address}`
-          : '#'
-      }
-      className={cx(styles.etherscanLink, styles.link)}
-    >
-      {label || children || addressShort}
-    </a>
-  )
-}
-
-export const Address = ({ isExchange, labels, ...rest }) => {
-  return (
-    <>
-      <EtherscanLink {...rest} isExchange={isExchange} />
-      {labels && <TransactionTableLabels labels={labels} />}
-    </>
-  )
 }
 
 export default WalletLink
