+++ src/ducks/HistoricalBalance/balanceView/utils.js
@@ -1,0 +1,3 @@
+export const getPriceMetricWithSlug = slug => {
+  return 'priceUsd@' + slug
+}
