+++ src/ducks/SANCharts/utils.js
@@ -1,6 +1,5 @@
 import React from 'react'
 import { YAxis, Bar, Line, Area } from 'recharts'
-import { getDateFormats, getTimeFormats } from '../../utils/dates'
 import { formatNumber, millify } from './../../utils/formatting'
 import ActiveLine from './tooltip/ActiveLine'
 import { Metric } from '../dataHub/metrics'
@@ -20,18 +19,6 @@ export const mapDatetimeToNumber = timeseries =>
     ...rest,
     datetime: +new Date(datetime)
   }))
-
-const currencyFormatter = (val, currency) =>
-  val || val === 0 ? formatNumber(val, { currency }) : 'No data'
-
-export const usdFormatter = val => currencyFormatter(val, 'USD')
-
-export const btcFormatter = val => currencyFormatter(val, 'BTC')
-
-export const ethFormatter = val => currencyFormatter(val, 'ETH')
-
-export const percentageFormatter = val =>
-  val || val === 0 ? `${val}%` : 'No data'
 
 const getEventColor = (isAnomaly, value) => {
   if (isAnomaly || value < 4) {
@@ -396,35 +383,3 @@ export const yAxisTickFormatter = value =>
     : formatNumber(value, {
       minimumFractionDigits: 0
     })
-
-export const tooltipValueFormatter = ({
-  value,
-  formatter,
-  threshold = 1000
-}) => {
-  try {
-    if (formatter) {
-      return formatter(value)
-    }
-
-    const numValue = +value
-    // NOTE(vanguard): Some values may not be present in a hovered data point, i.e. value === undefined/null;
-    if (!Number.isFinite(numValue)) throw new Error()
-
-    if (numValue > threshold) {
-      return millify(numValue, 2)
-    }
-
-    return numValue.toFixed(2)
-  } catch (e) {
-    return 'No data'
-  }
-}
-
-export const tooltipLabelFormatter = value => {
-  const date = new Date(value)
-  const { MMMM, DD, YYYY } = getDateFormats(date)
-  const { HH, mm } = getTimeFormats(date)
-
-  return `${HH}:${mm}, ${MMMM} ${DD}, ${YYYY}`
-}
