+++ src/ducks/Signals/chart/preview/utils.js
@@ -62,3 +62,14 @@ export const mapToRequestedMetrics = (
     address,
     ...reqMeta
   }))
+
+export const getAvailableCooldown = baseCooldown => {
+  if (
+    baseCooldown &&
+    (baseCooldown.indexOf('d') !== -1 || baseCooldown.indexOf('w') !== -1)
+  ) {
+    return '1d'
+  }
+
+  return baseCooldown && baseCooldown.indexOf('m') !== -1 ? '1h' : baseCooldown
+}
