+++ src/ducks/Signals/notifications/SignalNotificationActions.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import RemoveSignalButton from '../../../components/SignalCard/controls/RemoveSignalButton'
 import NotificationActions from '../../../components/NotificationActions/NotificationActions'
-import { canOpenTrigger } from '../../../components/SignalCard/card/SignalCard'
+import { canOpenTrigger } from '../../../components/SignalCard/card/utils'
 
 const SignalNotificationActions = ({ signal, toLink }) => {
   const { id } = signal
