+++ src/ducks/Signals/signalFormManager/sharedForm/SharedTriggerForm.js
@@ -7,7 +7,7 @@ import SignalCard from '../../../../components/SignalCard/card/SignalCard'
 import NoSignalPreview from '../../chart/preview/NoSignalPreview'
 import { DesktopOnly, MobileOnly } from '../../../../components/Responsive'
 import CopySignal from '../../../../components/SignalCard/controls/CopySignal'
-import { isStrictTrendingWords } from '../../../../components/SignalCard/card/TrendingWordsSignalCard'
+import { isStrictTrendingWords } from '../../../../components/SignalCard/card/utils'
 import styles from './ShareTriggerForm.module.scss'
 
 const SharedTriggerForm = ({
