+++ src/ducks/Studio/AdvancedView/PriceHistogram/index.js
@@ -7,12 +7,12 @@ import ErrorMessage from './ErrorMessage'
 import UsageTip from '../UsageTip'
 import Calendar from '../Calendar'
 import MoreInfoLink from '../../../../components/MoreInfoLink/MoreInfoLink'
-import { usdFormatter } from '../../../SANCharts/utils'
 import { millify } from '../../../../utils/formatting'
 import { ONE_MONTH_IN_MS } from '../../../../utils/dates'
-import styles from './index.module.scss'
 import MetricFrequence from '../../../SANCharts/MetricFrequence/MetricFrequence'
 import { SPENT_COINT_COST } from '../../../dataHub/metrics/frequences'
+import { usdFormatter } from '../../../dataHub/metrics/formatters'
+import styles from './index.module.scss'
 
 const INTERVAL_ERROR_TEXT = 'allowed interval'
