+++ src/ducks/Studio/Sidebar/index.js
@@ -6,9 +6,15 @@ import Icon from '@santiment-network/ui/Icon'
 import MetricSelector from './MetricSelector'
 import InsightAlertSelector from './InsightAlertSelector'
 import Search from './Search'
+import { HOLDER_DISTRIBUTION_NODE } from './nodes'
 import withMetrics from '../withMetrics'
+import { TopTransactionsTableMetric } from '../../dataHub/submetrics'
 import { rebuildDescriptions } from '../../dataHub/metrics/descriptions'
 import styles from './index.module.scss'
+
+const HOLDER_DISTRIBUTION_ITEM = { item: HOLDER_DISTRIBUTION_NODE }
+const TOP_TRANSACTIONS_TABLE_ITEM = { item: TopTransactionsTableMetric }
+const ON_CHAIN_DEFAULT = [HOLDER_DISTRIBUTION_ITEM, TOP_TRANSACTIONS_TABLE_ITEM]
 
 const TABS = ['Metrics', 'Insights']
 const DEFAULT_TAB = TABS[0]
@@ -28,7 +34,7 @@ const Header = ({ activeTab, setActiveTab, ...props }) => (
       // NOTE: Not passed as a reference, since more than 1 argument is passed to a callback [@vanguard | Aug  4, 2020]
       onSelect={tab => setActiveTab(tab)}
     />
-    <Search {...props} />
+    <Search onChainDefault={ON_CHAIN_DEFAULT} {...props} />
   </div>
 )
