+++ src/ducks/Studio/Sidebar/nodes.js
@@ -1,0 +1,8 @@
+import { WIDGET } from './Button/types'
+
+export const HOLDER_DISTRIBUTION_NODE = {
+  key: 'holder_distribution',
+  type: WIDGET,
+  label: 'Holder Distribution',
+  abbreviation: 'hd'
+}
