+++ src/ducks/Studio/withMetrics.js
@@ -1,9 +1,9 @@
 import gql from 'graphql-tag'
 import { graphql } from 'react-apollo'
 import { getCategoryGraph } from './Sidebar/utils'
+import { getMarketSegment } from './timeseries/marketSegments'
 import { useMergedTimeboundSubmetrics } from '../dataHub/timebounds'
 import { getAssetNewMetrics } from '../dataHub/metrics/news'
-import { getMarketSegment } from './timeseries/marketSegments'
 
 const PROJECT_METRICS_QUERIES_SEGMENTS_BY_SLUG_QUERY = gql`
   query projectBySlug($slug: String!) {
