+++ src/pages/Login/LoginEmailForm.js
@@ -7,7 +7,7 @@ import isEqual from 'lodash.isequal'
 import GA from '../../utils/tracking'
 import { InputWithIcon as Input } from '@santiment-network/ui/Input'
 import Button from '@santiment-network/ui/Button'
-import { PATHS } from '../../App'
+import { PATHS } from '../../paths'
 import MobileWrapper from './Mobile/MobileWrapper'
 import FormikInput from '../../components/formik-santiment-ui/FormikInput'
 import FormikEffect from '../../components/formik-santiment-ui/FormikEffect'
