+++ src/pages/Login/index.js
@@ -8,7 +8,7 @@ import LoginMetamaskBtn from './LoginMetamaskBtn'
 import LoginEmailForm from './LoginEmailForm'
 import LoginEmailBtn from './LoginEmailBtn'
 import FreeTrialBlock from './FreeTrialBlock'
-import { PATHS } from '../../App'
+import { PATHS } from '../../paths'
 import SwipablePages from '../../components/SwipablePages/SwipablePages'
 import MobileWrapper from './Mobile/MobileWrapper'
 import { LoginDivider } from './CreateAccountFreeTrial'
