+++ src/pages/feed/GeneralFeed/FeedList/dates.js
@@ -1,0 +1,2 @@
+export const TODAY = new Date().toLocaleDateString()
+export const YESTERDAY = addDays(new Date(), -1).toLocaleDateString()
