+++ src/paths.js
@@ -1,0 +1,12 @@
+export const PATHS = {
+  FEED: '/feed',
+  SOCIAL_TOOl: '/labs/trends/explore/',
+  LOGIN: '/login',
+  LOGIN_VIA_EMAIL: '/login/email',
+  CREATE_ACCOUNT: '/sign-up',
+  GDPR: '/gdpr',
+  PRO_METRICS: '/pro-sheets-templates',
+  INDEX: '/',
+  STUDIO: '/studio',
+  CHARTS: '/charts'
+}
